+++ test/e2e/alertmanager_test.go
@@ -23,7 +23,7 @@ import (
 	"k8s.io/client-go/pkg/api/v1"
 
 	"github.com/coreos/prometheus-operator/pkg/client/monitoring/v1alpha1"
-	testFramework "github.com/coreos/prometheus-operator/test/e2e/framework"
+	testFramework "github.com/coreos/prometheus-operator/test/framework"
 )
 
 func TestAlertmanagerCreateDeleteCluster(t *testing.T) {
