+++ test/framework/replication-controller.go
