+++ test/framework/ressources/nxginx-ingress-controller.yml
